Add a new option that will be used to test whether the compiler is
Estimated hours taken: 0.25
Branches: main

Add a new option that will be used to test whether the compiler is
recent enough to compile the changes to the library that require
foreign_import_module pragma to be handled correctly.

This will be needed by the configure test at a later date, when using a
foreign_type for io__stream is re-enabled.

compiler/options.m:
    Add the option bug-foreign_import-2002-08-06.