The "complete" and "incomplete" maps in the declarative debugging oracle
Estimated hours taken: 2
Branches: main

The "complete" and "incomplete" maps in the declarative debugging oracle
can cause a runtime abort if higher order terms are present in the debuggee,
because the maps use sets to represent the solutions to a goal, and sets make
use of builtin comparison.  This change fixes the problem by replacing the
two maps with a single map from initial atom to truth value.  The set of
solutions is no longer stored explicitly.

This change causes no loss of accuracy of the oracle, because we weren't
really using the explicit knowledge about solutions anyway.  The main
drawback of this change is that if in future we implement support for "edit
and continue" style debugging (e.g. by allowing previous answers to be
saved in/restored from a file), we will lose some of the information
pertaining to previous versions of the program, which will make the oracle
slightly less effective at deducing new answers from previous ones.  Since
runtime aborts are a much more serious problem, we will have to live with
this drawback for the moment.

Another solution to this problem may have been to implement a 'set_cc'
module, which would be to 'set' as 'tree234_cc' is to 'tree234'.
Unfortunately, we would need to implement 'set_cc__intersect' but it
appears not to be possible to implement a satisfactory committed choice
version of intersection.

browser/declarative_oracle.m:
	Replace the two maps with the new one, and update the code
	that accesses the knowledge base.

tests/debugger/declarative/Mmakefile:
	Enable tests 'ho3' and 'ho4', which now work.

tests/debugger/declarative/ho3.m:
	Fix a typo in the test case.  It wasn't testing precisely the thing
	it was supposed to test.

tests/debugger/declarative/ho3.exp:
tests/debugger/declarative/ho3.inp:
tests/debugger/declarative/ho4.exp:
tests/debugger/declarative/ho4.inp:
	Input and expected output for the enabled tests.