Change the way we attach the call_table_gen feature to the goal that computes
Estimated hours taken: 3
Branches: main

Change the way we attach the call_table_gen feature to the goal that computes
the node at the tip of the call table. We still generate a conjunction
of lookup goals, one for each input argument. However, instead of
attaching the feature to the conjunction and wrapping the conjunction
in a cannot_remove some() goal to protect the conjunction against flattening
(since flattening would remove the feature), we now simply attach the feature
to the rightmost lookup goal, the one that actually computes the call table
tip node. Since this goal has an output that is needed, it cannot be optimized
away, and optimizations  cannot transform it either, because it is a primitive
(a call to a procedure designed by foreign_proc).

Getting rid of the some() goals avoids redundant pushes and pops on the cut
stack in minimal model grades.

compiler/table_gen.m:
	Change the tabling transformation along the lines described above.

tests/debugger/print_table.exp:
tests/debugger/retry.exp:
	Update the expected output to show goal paths in which the conjunction
	of lookup goals is flattened out.