Add pragma terminates declarations to exception.builtin_throw/1,
Estimated hours taken: 1
Branches: main

Add pragma terminates declarations to exception.builtin_throw/1,
exception.throw/1 and require.error/1.

The point of this change is to try and reduce the number of cases of
possible non-termination being reported in the standard library due to
calls to require.error/1 or exception.throw/1.

Ideally we would only need provide a terminates declaration for
builtin_throw/1 but in practice the structure of the module dependency graph
means that error/1 is marked as non-terminating and certain calls to throw/1
are also considered to be non-terminating.

library/exception.m:
	Add pragma terminates declaration for builtin_throw/1 and
	throw/1.
library/require.m:
	Add pragma terminates declaration for error/1.