Fix a bug introduced by petdr's Dec 2 bug fix which introduced wrapper
Estimated hours taken: 3
Branches: main

Fix a bug introduced by petdr's Dec 2 bug fix which introduced wrapper
functions for unify and compare procs to unbox their arguments,
which caused the tests/valid/unify_typeinfo_bug test case to fail.

The bug was that with --no-special-preds, we were generating invalid
MLDS code, because the code to introduce wrapper functions assumed
that the function being wrapped was defined in the same module,
but with --no-special-preds, the procedure being wrapped is always
private_builtin.unused/0.

In the case of a procedure with arity zero, the wrapper is not needed
(it's just a tail call), so the simplest fix was to just optimize
that case.

compiler/rtti_to_mlds.m:
	If the procedure being wrapped has arity zero, don't bother to
	generate a wrapper function, since it is not needed.