Reduce the size of the generated C source files in debugging grades.
Estimated hours taken: 6
Branches: main

Reduce the size of the generated C source files in debugging grades. The
motivation for this change is the observation that a workspace compiled in
a debug grade takes up a large fraction of a gigabyte, and the largest
fraction of this is the C files. The largest debug-grade generated source
file, parse_tree.modules.c, is 23.5 megabytes (!) in size, whereas
parse_tree.modules.o is only about 3.5 megabytes (modules.m itself is only
about a quarter of a megabyte).

We use three techniques to reduce the C source file size.

The first is to get rid of unnecessary casts of constants in data structure
initialization. When putting an integer constant into a field of type (e.g.)
MR_uint_least8_t, the general purpose rval output routines we used to use
cast the constant first to MR_Integer and then to MR_uint_least8_t, even
though in that specific case neither cast is needed. Similarly, there is no
need to cast pointer constants (or the integer 0) twice when used in
contexts where pointers are expected; one cast to the final type is enough.

The second technique is to take advantage of sequences of fields of the
same type. Instead of using long type definitions with one field per argument,
we can now use much shorter type definitions with one field, an array,
per run of arguments of the same type.

The third technique is reduce the length of the names of the data structures
containing common cells and their types, by reducing the length of their
prefixes and module qualify them only if we have to, and deleting the duplicate
mercury_ prefixes from the names of some debugger data structures.

Together, these changes reduce the size of parse_tree.modules.c to 13.2 Mb,
a reduction of more than 43%. There is a small reduction in the size of the
object file (about 3%, due to shorter symbol names in the relocation section),
and there is a very slight increase in compilation time in debug grades
(in non-debug grades, the increase in compilation time is negligible).
Given the inconvenience of running out of disk space, this is a good tradeoff.

compiler/llds.m:
	Change the definition of static term data structures to allow for
	the type of the cell being defined in terms of structure fields
	containing arrays.

compiler/llds_out.m:
	Change the code that writes out the types and the contents of
	static term data structures to handle structure fields containing
	arrays.

	Do not cast integer initializers unnecessarily.

	Do not module qualify the names of the global variables holding
	common cells unnecessarily, and use shorter prefixes for them.

compiler/global_data.m:
	Look for opportunities to group sequences of fields into arrays.

	Use the more compact description of cell types possible when using
	arrays to speed up lookups in the cell type map.

	Change a predicate interface to make it cheaper to implement in the
	presence of array fields.

compiler/var_locn.m:
	Use the updated interface in global_data.m.

compiler/name_mangle.m:
	Provide functions for the shorter prefixes for common cell names and
	types.

	Provide a predicate for printing proc_labels without the mercury_
	prefix.

compiler/layout_out.m:
	Omit the duplicate mercury_ prefixes within the names of most debugger
	data structures. (Some need it to allow C macros to derive their names
	from the names of labels.)