Fix problems in white space in the commit of the fix of the deep profiler.
Estimated hours taken: 0.2
Branches: main

Fix problems in white space in the commit of the fix of the deep profiler.
These were caused by the fact that the workspace in which the original commit
took place was created by applying a patch to a fresh workspace, and the patch
was created with diff -b.