Some of the tabling functions in the runtime are incompatible
Estimated hours taken: 2
Branches: main

Some of the tabling functions in the runtime are incompatible
with the library when using the hl*_nest* grades.  The following
diff fixes this to the extent that the library compiles
and passes most of the tests.

runtime/mercury_tabling.h:
runtime/mercury_tabling.c:
	Handle memoing functions when using compiling
	using the hl*_nest* grades.