Fix the failures of the tabling/loopcheck_nondet test case at optimization
Estimated hours taken: 6
Branches: main

Fix the failures of the tabling/loopcheck_nondet test case at optimization
levels that enabled --local-constraint-propagation. The problem was that
constraint propagation reran determinism analysis. This changed the
determinism of an impure disjunction from multi to det. This prevented
backtracking into the second disjunct, and this in turn caused the
transformation not to detect the infinite loop in the program.

compiler/table_gen.m:
	Attach the preserve_backtrack_info goal feature to all goals of
	determinism failure we construct, since we don't construct any goals
	we don't mean to execute.

	Also add some missing variables to nonlocal sets in goal_infos.

compiler/constraint.m:
	Preserve the context and all features of goals as far as possible.
	Before this change, the constraint propagation transformation deleted
	the hide_debug_event annotations from some goals, which could alter
	what events an mdb user could see with --trace-optimized.

compiler/pd_debug.m:
	Fix a bug in my previous cleanup that caused an infinite loop when
	using --debug-pd.