Add the option `--file' to mmake.  This allows the user
Estimated hours taken: 1
Branches: main

Add the option `--file' to mmake.  This allows the user
to specify the name of the file that mmake uses as the Mmakefile.
This can be useful when porting programs between systems
and experimenting with different configurations, particularly
when the programs are not under the control of autoconf.

The option `--file' may be abbreviated to `-f'.  If a file
is specified via `--file' then it will be used in preference to
any file in the current directory named Mmakefile or Mmake.

Unlike GNU make using multiple instances of `--file' will
not cause mmake to jointly use all of the specified files.
Instead it will use the last one specified and ignore the
others.

scripts/mmake.in:
	Add the `--file' option to mmake.