This is phase 1 of a 2 phase diff to get Mercury shared libs working on Mac OS
Estimated hours taken: 20
Branches: main

This is phase 1 of a 2 phase diff to get Mercury shared libs working on Mac OS
X.  In this phase support for three new compiler options (documented below) are
added.  In the next phase changes to the configure scripts that use these
options will be committed.

Darwin shared libraries have the extension `dylib' instead of `so'.  Also
objects that link to a shared library on Darwin don't need to be told the
runtime path in which they should look for the libraries, instead the
shared libraries themselves remember where they will be eventually installed
(called the install-name) and any object which links in the shared library will
get the install-name from the shared library at link time.  When a shared
library is built it has to be told where it will be installed which is what
the libtool -install_name option is used for on Darwin.

This diff only enables shared libraries on Darwin, not "bundles" which are
shared objects that can be loaded dynamically at runtime using something like
dlopen.  Therefore the interactive query tool in the debugger still doesn't
work on Mac OS X.

Added three new compiler options :

--shlib-linker-use-install-name :
	A boolean flag to tell the compiler to use the -install_name option
	when building shared libraries.  When this flag is set the following
	options have no effect: --linker-rpath-flag, --linker-rpath-separator,
	--shlib-linker-rpath-flag, --shlib-linker-rpath-separator.

--shlib-linker-install-name-flag :
	The flag name to use ("-install_name" for Darwin).

--shlib-linker-install-name-path :
	The path where the shared library will eventually end up, excluding the
	file name.  The file name is appended to the end before the option
	is passed on to the linker.

compiler/compile_target_code.m
	Passed the install-name options to the linker and made RpathsOpts empty
	if the --shlib-linker-use-install-name flag is set.

compiler/modules.m
	Added get_install_name_option/4 predicate to get the option string to
	set the install-name for a shared library.

compiler/options.m
	Added --shlib-linker-install-name-path,
	--shlib-linker-install-name-flag and --shlib-linker-use-install-name
	options.