Set default value of --shlib-linker-install-name-path back to the value of
Estimated hours taken: 0.1
Branches: main

Set default value of --shlib-linker-install-name-path back to the value of
INSTALL_MERC_LIB_DIR, so building of user shared libraries works as described
in the user guide.  This was set to FINAL_INSTALL_MERC_LIB_DIR previously so
that Mercury distribution packages (in particular fink) could be built
correctly, but this is not a problem anymore since
--shlib-linker-install-name-path is now set manually in the Mmakefiles for the
Mercury build.

compiler/options.m
	Changed default value of --shlib-linker-install-name-path.