Fix a code generator bug that manifested itself in profiling grades with
Estimated hours taken: 3
Branches: main

Fix a code generator bug that manifested itself in profiling grades with
(intermodule) inlining, the symptom being a compiler abort due to a stackvar
out of range (negative slot number). The problem was that we use the type of a
variable to decide whether the variable has a real stack slot or not, but
we were using the original type of an argument of a foreign_proc to decide
whether the argument should be passed or not. Inlining could cause an argument
whose original type is T (e.g. the arguments of cc_multi_equal) to be
instantiated, e.g. to io__state. Since the type of the variable is io__state,
it gets a dummy stack slot; since the original type of the argument is T,
we pass the variable and thus refer to the dummy stack slot.

The fix is the use the actual type of the variable to make the decision,
not the original type of the argument.

This fix fixes the failures of the deforest_cc_bug.m test case in deep
profiling grades and the tricky_try_store.m test case in profiling grades.

compiler/llds.m:
	Add an extra argument to the pragma_c_{in,out}put function symbols
	to hold the actual type of the variable being passed.

	Add field names to those function symbols.

compiler/pragma_c_gen.m:
	Fill in the new argument.

	Clean up some code, mostly by factoring out duplicate code fragments.

compiler/llds_out.m:
	Handle the new argument to decide whether to pass the argument.

compiler/exprn_aux.m:
compiler/livemap.m:
compiler/middle_rec.m:
compiler/opt_util.m:
	Handle the new argument in pragma_c_{in,out}puts.