Improve the interface of exception__try_all by separating the output into
Estimated hours taken: 1
Branches: main

Improve the interface of exception__try_all by separating the output into
a maybe-exception and a list of solutions.  This makes it easier to test
whether an exception occurred, and removes the need for solutions to be
wrapped in an exception_result.

This also obviates a bug in the previous implementation/documentation
of try_all whereby it was putting the exception at the start of the results
list instead of the end, as documented.

Although this change is not backwards compatible, the fact that no one
reported the existing bug for the last four years suggests that that
shouldn't be much of an issue.

NEWS:
	Mention the change.

library/exception.m:
	Add try_all/3, which is like try_all/2 except as described above.

	Make try_all/2 obsolete.  Modify its documentation to reflect its
	actual behaviour.  Implement it in terms of try_all/3.

	Rewrite the documentation of incremental_try_all so as not to refer
	to try_all/2.  The meaning of incremental_try_all is not changed.

tests/hard_coded/exceptions/test_try_all.{m,exp}:
	Update the test case to reflect the new interface.