This change adds a --untuple option to the compiler which invokes a new
Estimated hours taken: 80
Branches: main

This change adds a --untuple option to the compiler which invokes a new
transformation pass.  The transformation takes the HLDS structure as its
input and transforms the locally-defined procedures as follows: if the
formal parameter of a procedure has a type consisting of a single function
symbol then that parameter is expanded into multiple parameters (one for
each field of the functor).  Tuple types are also expanded.  The argument
lists are expanded as deeply (flatly) as possible.  Calls within the
same module are updated to use the new versions of the procedures.

It is expected that applying the transformation results in a performance
degradation for almost all cases.  However, it will be useful for the
development of a related tupling transformation (underway) and may also
be useful in conjunction with it.

compiler/untupling.m:
	New file.

compiler/transform_hlds.m:
	Add the new untupling submodule to the transform_hlds module.

compiler/handle_options.m:
compiler/options.m:
	Add a new boolean option --untuple to the compiler.

compiler/mercury_compile.m:
	Add code to run the untupling pass if the --untuple option is given.

compiler/hlds_pred.m:
compiler/layout_out.m:
	Add a untuple constructor to the pred_transformation type.

	Add a untuple_proc_info field to proc_sub_info and associated
	predicates proc_info_get_maybe_untuple_info and
	proc_info_set_maybe_untuple_info.

compiler/hlds_out.m:
	Add code to print out the information stored in the
	untuple_proc_info field of proc_sub_info.

compiler/hlds_goal.m:
	Add new utility predicates construct_functor and deconstruct_functor
	for creating construct and deconstruct unification goals.

	Make construct_tuple and deconstruct_tuple be special cases of
	construct_functor and deconstruct_functor, respectively.