Don't run tests/hard_coded/parse on OS X as it is
Estimated hours taken: 0.1
Branches: main

Don't run tests/hard_coded/parse on OS X as it is
statically linked, not only with the Mercury libraries,
but also the system ones.  This doesn't work on OS X
because the static versions of the system libraries
are not installed by default.

tests/hard_coded/Mmakefile:
	Disable tests/hard_coded/parse on OS X.