Fix a bug that has been causing the following tests to
Estimated hours taken: 2
Branches: main, release

Fix a bug that has been causing the following tests to
fail on aral in the .mmsc grades.

	grade_subdirs/hello
	mmc_make/hello
	mmc_make/rebuild

The problem was that the grade string returned by
grade_directory_component/2 would contain not only the
.mmsc component but also the .mm  (which is an alternative
name for it).  This was a problem for mmc --make because it
ended up looking for the library in directory that did not
exist.

compiler/handle_options.m:
	In .mmsc and .dmmsc grades, make sure that the
	grade strings returned by grade_directory_component/2
	do not also contain the components .mm and .dmm.

	Fix a typo in a comment:s/procesing/processing/