Fix tests/valid/untuple_bug so that it does not fail in the .decldebug
Estimated hours taken: 0.1
Branches: main

Fix tests/valid/untuple_bug so that it does not fail in the .decldebug
grades.  The problem is that is compiled with `--trace deep' which is
incompatible with the .decldebug grades.

The fix is to compile the test case with `--trace rep'.  The original
bug still shows up with trace rep if Peter's fix is removed.

tests/valid/Mercury.options:
	Compile the untuple_bug test case with `--trace rep'
	rather than `--trace deep'.