Merge some of the recent changes to the mdbcomp library
Estimated hours taken: 1
Branches: release

Merge some of the recent changes to the mdbcomp library
from the main branch onto the release branch.  The reason
for doing this is that one of the recent changes to solver
types that was committed on the release branch used features
of mdbcomp that were only found on the main branch.

In preference to doing anything else I've updated mdbcomp
(and the relevant compiler modules).  Hopefully this should
make adding things to both branches a bit easier in the future.

mdbcomp/prim_data.m:
	Merge in changes from the main branch.

compiler/rl_out.pp:
compiler/*.m:
	Update as necessary.