Reduce the size of the bytecode representations of procedures (generated with
Estimated hours taken: 10
Branches: main

Reduce the size of the bytecode representations of procedures (generated with
the --trace rep option) by applying two optimizations.

The first optimization is to store strings in the existing global string table
and just keep the offsets in the bytecode representation.  This reduces the
size of the procedure representation bytecode for mercury_compile by 15.46% and
the total size of the executable by 2.34%.

The second optimization is to use a single byte to store variable numbers
where the biggest variable number in a procedure is less than or equal to 255.
An extra byte is stored at the beginning of the procedure to indicate whether
variable numbers are stored as bytes or shorts.  This optimization reduces the
size of the procedure representation bytecode for mercury_compile by 11.34% and
the total size of the executable by 1.75%.

browser/declarative_execution.m:
	Pass the label_layout for a call to any predicates than need to
	look up strings in the bytecode.  The global string table is accessable
	via the label_layout.

	Pass the representation used for variable numbers (byte or short) to
	any predicates that need to look up variable numbers in the bytecode.

compiler/c_util.m:
	Escape (') characters in the C representation of the string table.

compiler/prog_rep.m:
	Update the global string table when converting a string to bytecode.

	Determine the maximum variable number in a procedure and use a byte
	to represent the variable numbers if this is less than 256.

compiler/stack_layout.m:
	Export the predicate that adds a string to the global string table.
	Export an abstract version of the stack_layout_info type that the
	above predicate operates on.

	Pass the stack_layout_info to the predicate that constructs the
	bytecodes for a procedure, so it can update the global string
	table.

mdbcomp/program_representation.m:
	Add a predicate to convert the byte indicating whether bytes or shorts
	are used to represent variable numbers to a Mercury value.

trace/mercury_trace_internal.c:
	Pass the label layout when reading a procedure from the bytecode, so
	that strings can be looked up in the global string table.