When sourcing a file of mdb commands, add the commands to the front of the
Estimated hours taken: 2
Branches: main

When sourcing a file of mdb commands, add the commands to the front of the
queue, while maintaining their original order.

This has the effect replacing `source' commands with the
commands in the sourced file.  Previously sourced commands would be added
to the end of the queue, so they would appear after any commands that might
appear after the original `source' command.

For example if .mdbrc was set to:

break p
continue
finish
dd

then a whole lot of error messages would be displayed, because the commands
to read in the documentation would be executed after the `dd' command above
(and the declarative debugger doesn't recognise those commands).  Now the
commands to read in the documentation are executed first.

Also include the `condition' command in the list of breakpoint commands.

doc/mdb_categories:
	Include `condition' in the list of breakpoint commands.

trace/mercury_trace_internal.c:
	When sourcing commands, insert them at the front of the queue.  This
	requires the calls to the functions that read the default source
	files to be reversed.